fix element update logic
